refactor(foundation): split CarModeView, Globe, variants, and docs SSOT
- Extract CarModeView controllers into src/views/car/* hooks and overlay components (313 LOC)
- Add Cesium type facade (src/types/cesium.ts) and globe POI/autopilot modules
- Split ScienceLabMode geometry/materials/audio; extract Convertible WindParticleSystem
- Share MobileUI panel contracts via chromePanelContracts with ConnectedChrome
- Refresh DEVELOPER_CONTEXT, README TS badge, archive stale plan.md checkboxes
