+++ src/index.tsx
@@ -1,3 +1,4 @@
+import './types/cesium';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
